Extract getName function to a separate file
