+++ react/ShippingSummary.tsx
@@ -4,8 +4,10 @@ import { AddressContext } from 'vtex.address-context'
 import { OrderShipping } from 'vtex.order-shipping'
 import { TranslateEstimate } from 'vtex.shipping-estimate-translator'
 import { FormattedCurrency } from 'vtex.format-currency'
+import { DeliveryOption, PickupOption } from 'vtex.checkout-graphql'
 
 import useAddressRules from './useAddressRules'
+import { getName } from './utils/sla'
 
 const { useOrderShipping } = OrderShipping
 
@@ -38,17 +40,19 @@ const ShippingSummary: React.FC = () => {
       >
         <PlaceDetails display="extended" />
       </AddressContext.AddressContextProvider>
-      {selectedShippingOptions.map((shippingOption: any) => (
-        <div className="mt5" key={shippingOption.id}>
-          <span>
-            {shippingOption.id} &ndash; {''}
-            <FormattedCurrency value={shippingOption.price / 100} />
-          </span>
-          <p className="mv0">
-            <TranslateEstimate shippingEstimate={shippingOption.estimate} />
-          </p>
-        </div>
-      ))}
+      {selectedShippingOptions.map(
+        (shippingOption: DeliveryOption | PickupOption) => (
+          <div className="mt5" key={shippingOption.id}>
+            <span>
+              {getName(shippingOption)} &ndash; {''}
+              <FormattedCurrency value={shippingOption.price / 100} />
+            </span>
+            <p className="mv0">
+              <TranslateEstimate shippingEstimate={shippingOption.estimate} />
+            </p>
+          </div>
+        )
+      )}
     </div>
   )
 }
