+++ react/utils/sla.ts
@@ -1,0 +1,7 @@
+import { DeliveryOption, PickupOption } from 'vtex.checkout-graphql'
+
+export const getName = (shippingOption: DeliveryOption | PickupOption) => {
+  return shippingOption.deliveryChannel === 'pickup-in-point'
+    ? (shippingOption as PickupOption).friendlyName
+    : shippingOption.id
+}
